refactor: input, output view 분리
