+++ src/main/java/RacingMain.java
@@ -1,13 +1,13 @@
 import controller.RacingGameController;
 import controller.dto.CarNamesRequest;
 import controller.dto.TryCountRequest;
-import view.InputOutputView;
+import view.InputView;
 
 public class RacingMain {
 
     public static void main(final String[] args) {
-        final String carNames = InputOutputView.getCarNames();
-        final String tryCount = InputOutputView.getTryCount();
+        final String carNames = InputView.getCarNames();
+        final String tryCount = InputView.getTryCount();
 
         RacingGameController gameController = RacingGameController.of(
             CarNamesRequest.from(carNames),
